Remove bad comments and add some more constructors

+++ android/src/main/java/net/daverix/slingerorm/android/SQLiteDatabaseConnection.java
@@ -4,6 +4,7 @@ import android.content.ContentValues;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import net.daverix.slingerorm.DatabaseConnection;
+import net.daverix.slingerorm.android.internal.CursorResultsFactory;
 import net.daverix.slingerorm.android.internal.InsertableContentValues;
 import net.daverix.slingerorm.android.internal.ResultRowsFactory;
 import net.daverix.slingerorm.mapping.InsertableValues;
@@ -16,12 +17,23 @@ import javax.inject.Inject;
  */
 public class SQLiteDatabaseConnection implements DatabaseConnection {
     private final SQLiteDatabase mDb;
+
     private final ResultRowsFactory mResultRowsFactory;
 
     @Inject
     public SQLiteDatabaseConnection(SQLiteDatabase db, ResultRowsFactory resultRowsFactory) {
+        if(db == null) throw new IllegalArgumentException("db is null");
+        if(resultRowsFactory == null) throw new IllegalArgumentException("resultRowsFactory is null");
+
         mDb = db;
         mResultRowsFactory = resultRowsFactory;
+    }
+
+    public SQLiteDatabaseConnection(SQLiteDatabase db) {
+        if(db == null) throw new IllegalArgumentException("db is null");
+
+        mDb = db;
+        mResultRowsFactory = new CursorResultsFactory();
     }
 
     @Override
@@ -41,11 +53,15 @@ public class SQLiteDatabaseConnection implements DatabaseConnection {
 
     @Override
     public void execSql(String sql) {
+        if(sql == null) throw new IllegalArgumentException("sql is null");
+
         mDb.execSQL(sql);
     }
 
     @Override
     public void execSql(String sql, String[] args) {
+        if(sql == null) throw new IllegalArgumentException("sql is null");
+
         mDb.execSQL(sql, args);
     }
 
@@ -56,26 +72,39 @@ public class SQLiteDatabaseConnection implements DatabaseConnection {
 
     @Override
     public boolean replace(String tableName, InsertableValues values) {
+        if(tableName == null) throw new IllegalArgumentException("tableName is null");
+        if(values == null) throw new IllegalArgumentException("values is null");
+
         return mDb.replaceOrThrow(tableName, null, (ContentValues) values.getData()) != -1;
     }
 
     @Override
     public boolean insert(String tableName, InsertableValues values) {
+        if(tableName == null) throw new IllegalArgumentException("tableName is null");
+        if(values == null) throw new IllegalArgumentException("values is null");
+
         return mDb.insertOrThrow(tableName, null, (ContentValues) values.getData()) != -1;
     }
 
     @Override
     public int update(String tableName, InsertableValues values, String selection, String[] selectionArgs) {
+        if(tableName == null) throw new IllegalArgumentException("tableName is null");
+        if(values == null) throw new IllegalArgumentException("values is null");
+
         return mDb.update(tableName, (ContentValues) values.getData(), selection, selectionArgs);
     }
 
     @Override
     public int delete(String tableName, String selection, String[] selectionArgs) {
+        if(tableName == null) throw new IllegalArgumentException("tableName is null");
+
         return mDb.delete(tableName, selection, selectionArgs);
     }
 
     @Override
     public ResultRows query(boolean distinct, String tableName, String[] fields, String selection, String[] selectionArgs, String having, String groupBy, String orderBy) {
+        if(tableName == null) throw new IllegalArgumentException("tableName is null");
+
         Cursor cursor = mDb.query(distinct, tableName, fields, selection, selectionArgs, groupBy, having, orderBy, null);
         return mResultRowsFactory.create(cursor);
     }
